Legal_Assist_ML/requirements.txt
flask
flask-cors
pymongo
gensim
nltk
requests
python-dotenv
scikit-learn
